cmn_cyprus: Add cmn_hns_cfg_ctl to HN-S register layout
cmn_hns_cfg_ctl is the register functioning as the configuration
control register for HN-F(S). It contains several fields that
related to the MPAM system level cache configurations.

For the structure defining the register layout of HNS, cmn_hns_cfg_ctl
is included within the "RESERVED2" region. Define it as a separate
register. Additionally, rename the RESERVED regions as a new split is
introduced.

Upstream-Status: Pending
Suggested-by: Rohit Mathew <rohit.mathew@arm.com>
Signed-off-by: Henry Wang <Henry.Wang@arm.com>
Signed-off-by: Rohit Mathew <rohit.mathew@arm.com>
Signed-off-by: Wei Chen <wei.chen@arm.com>