product: Add Kronos support
RD-Kronos SCP is almost the same as RD-Fremont. In this case, we
use the RD-Fremont's RD-INFRA-2023.09.28 PLATFORM_VARIANT = 0 code
as the base code of RD-Kronos init support.

So in this patch, we duplicate the rdfremont product directory to
rdkronos, and remove the PLATFORM_VARIANT=[1|2] related codes, and
rename fremont to kronos.

Upstream-Status: Pending
Signed-off-by: Wei Chen <wei.chen@arm.com>