product/rdkronos: Add AP and SI shared SRAM to CMN memory region map
To improve FFI and determinism, in recent update of the Kronos, we moved
the shared memory between SI and Primary Compute from primary compute
to safety island subsystem.

We created a dedicated SRAM region in safety island and added an extra
HN-I on CMN-Cyprus for AP cores in Primary compute to access the memory
via Safety Island interconnect.

In this patch we add a new entry in the Kronos' cmn_cyprus memory region
map to cover the address region of the new HN-I as described in the spec.

Upstream-Status: Pending
Signed-off-by: Wei Chen <wei.chen@arm.com>